@@ -62,33 +62,30 @@
     "ext-tokenizer": "*",
     "ext-intl": "*",
     "ext-xml": "*",
-    "ext-ldap": "*",
     "ext-zip": "*",
     "ext-json": "*",
     "ext-fileinfo": "*",
     "ext-openssl": "*",
     "ext-simplexml": "*",
-    "adldap2/adldap2-laravel": "^5.0",
-    "bacon/bacon-qr-code": "1.*",
+    "adldap2/adldap2-laravel": "5.*",
+    "bacon/bacon-qr-code": "2.*",
     "bunq/sdk_php": "dev-master",
-    "danhunsaker/laravel-flysystem-others": "^1.3",
+    "danhunsaker/laravel-flysystem-others": "1.*",
     "davejamesmiller/laravel-breadcrumbs": "5.*",
     "doctrine/dbal": "2.*",
     "fideloper/proxy": "4.*",
-    "laravel/framework": "5.7.*",
-    "laravel/passport": "^7.0",
-    "laravelcollective/html": "5.7.*",
+    "laravel/framework": "5.8.*",
+    "laravel/passport": "7.*",
+    "laravelcollective/html": "5.8.*",
     "league/commonmark": "0.*",
     "league/csv": "9.*",
-    "league/flysystem-replicate-adapter": "^1.0",
-    "league/flysystem-sftp": "^1.0",
-    "league/fractal": "^0.17.0",
-    "litipk/flysystem-fallback-adapter": "0.1.2",
-    "mschindler83/fints-hbci-php": "^1.0",
-    "pragmarx/google2fa": "3.*",
+    "league/flysystem-replicate-adapter": "1.*",
+    "league/flysystem-sftp": "1.*",
+    "league/fractal": "0.*",
+    "litipk/flysystem-fallback-adapter": "0.*",
+    "mschindler83/fints-hbci-php": "1.*",
     "pragmarx/google2fa-laravel": "0.*",
-    "rcrowe/twigbridge": "0.9.*",
-    "twig/twig": "1.*"
+    "rcrowe/twigbridge": "0.9.*"
   },
   "require-dev": {
     "barryvdh/laravel-ide-helper": "2.*",
@@ -96,8 +93,7 @@
     "fzaninotto/faker": "1.*",
     "johnkary/phpunit-speedtrap": "^3.0",
     "mockery/mockery": "1.*",
-    "php-coveralls/php-coveralls": "^2.0",
-    "phpunit/phpunit": "7.*",
+    "phpunit/phpunit": "8.*",
     "roave/security-advisories": "dev-master"
   },
   "suggest": {
