@@ -1,39 +1,26 @@
 <?php
 
-/**
- * twigbridge.php
- * Copyright (c) 2018 thegrumpydictator@gmail.com
- *
- * This file is part of Firefly III.
- *
- * Firefly III is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Firefly III is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Firefly III. If not, see <http://www.gnu.org/licenses/>.
- */
-
-declare(strict_types=1);
-
-use TwigBridge\Extension\Laravel\Auth;
-use TwigBridge\Extension\Laravel\Config;
-use TwigBridge\Extension\Laravel\Dump;
-use TwigBridge\Extension\Laravel\Input;
-use TwigBridge\Extension\Laravel\Session;
+use TwigBridge\Extension\Laravel\Url;
 use TwigBridge\Extension\Laravel\Str;
 use TwigBridge\Extension\Laravel\Translator;
-use TwigBridge\Extension\Laravel\Url;
-use TwigBridge\Extension\Loader\Facades;
+use TwigBridge\Extension\Laravel\Session;
+use TwigBridge\Extension\Laravel\Dump;
+use TwigBridge\Extension\Laravel\Input;
+use TwigBridge\Extension\Laravel\Config;
+use TwigBridge\Extension\Laravel\Auth;
 use TwigBridge\Extension\Loader\Filters;
 use TwigBridge\Extension\Loader\Functions;
+use TwigBridge\Extension\Loader\Facades;
 use TwigBridge\Twig\Template;
+
+/**
+ * This file is part of the TwigBridge package.
+ *
+ * @copyright Robert Crowe <hello@vivalacrowe.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 /**
  * Configuration options for Twig.
@@ -67,6 +54,7 @@ return [
             'debug'               => env('APP_DEBUG', false),
 
             // The charset used by the templates.
+            // default: utf-8
             'charset'             => 'utf-8',
 
             // The base template class to use for generated templates.
@@ -119,7 +107,7 @@ return [
         |
         | Enabled extensions.
         |
-        | `Twig_Extension_Debug` is enabled automatically if twig.debug is TRUE.
+        | `Twig\Extension\DebugExtension` is enabled automatically if twig.debug is TRUE.
         |
         */
         'enabled'   => [
@@ -202,7 +190,7 @@ return [
         | Available functions. Access like `{{ secure_url(...) }}`.
         |
         | Each function can take an optional array of options. These options are
-        | passed directly to `Twig_SimpleFunction`.
+        | passed directly to `Twig\TwigFunction`.
         |
         | So for example, to mark a function as safe you can do the following:
         |
@@ -226,7 +214,7 @@ return [
             'elixir',
             'head',
             'last',
-            'old',
+            'mix',
         ],
 
         /*
@@ -237,7 +225,7 @@ return [
         | Available filters. Access like `{{ variable|filter }}`.
         |
         | Each filter can take an optional array of options. These options are
-        | passed directly to `Twig_SimpleFilter`.
+        | passed directly to `Twig\TwigFilter`.
         |
         | So for example, to mark a filter as safe you can do the following:
         |
